[cleanup] computedLineHeightForFontUnits should not (implicit) floor https://bugs.webkit.org/show_bug.cgi?id=304445
Reviewed by Antti Koivisto.

* Source/WebCore/css/CSSToLengthConversionData.cpp:
(WebCore::CSSToLengthConversionData::computedLineHeightForFontUnits const):
* Source/WebCore/css/CSSToLengthConversionData.h:

Canonical link: https://commits.webkit.org/304747@main

+++ Source/WebCore/css/CSSToLengthConversionData.h
@@ -74,7 +74,7 @@ public:
     const Element* elementForContainerUnitResolution() const { return m_elementForContainerUnitResolution.get(); }
 
     const FontCascade& fontCascadeForFontUnits() const;
-    int computedLineHeightForFontUnits() const;
+    float computedLineHeightForFontUnits() const;
 
     FloatSize defaultViewportFactor() const;
     FloatSize smallViewportFactor() const;
